Ajouter un participant ok mais §§§§ MANQUE cLE SECONDAIRE VERS dOSSIERrESERVATIONiD

+++ BoVoyageMetier/Entities/Participant.cs
@@ -12,7 +12,7 @@ namespace BoVoyageMetier.Entities
     public class Participant:Personne
     {
         public int NumeroUnique { get; set; }
-        public decimal Reduction { get; set; }
+        public double Reduction { get; set; }
 
         public int DossierReservationId { get; set; }
